Fix for RT-20880: IllegalStateException thrown from Service#restart().
Under some circumstances (such as if the task had already run) cancel returns false and the state is not changed to CANCELLED. Due to concurrency, the bindings & state of the service have not completed by the time reset is called by restart, and it is in the wrong state and throws an exception. Test added.